775ec7b4 Merge topic 'fix_empty_selection' into master
cd34e287 Merge topic 'ignore-constant-expr-msvc' into master
c6200b34 Merge topic 'fix_append_filter_crash' into master
d7aca982 Fix handling of empty selection on ParallelCoordinates.
f8430aef vtkSmartPyObject: ignore warning C4127 on MSVC
b673e361 cmake: remove unnecessary condition copies
ac920d1f Fixed logic to work when an empty vtkSelection is passed.
ab0254aa Fixed error in appending cell data
13450fe7 Merge topic 'bump-metaio' into master
c4e2f41c Merge remote-tracking branch 'pvvtk/pv-master' into HEAD
6a1d3bc4 Merge topic 'fix_scalar_bar_representation' into master
cd2415bd Merge topic 'vtkSmartPyObject_documentation' into master
befcfd1f Format comments according to VTK stlye guide
47539122 Merge topic 'python_syntax_highlighting' into master
c3ee6097 Merge topic 'fix-mpi4py-include-dir' into master
cbd95551 Merge topic 'dashboard-warnings-20140807' into master
...

+++ Rendering/ContextOpenGL/vtkOpenGLContextDevice2D.cxx
@@ -420,7 +420,7 @@ void vtkOpenGLContextDevice2D::DrawMarkers(int shape, bool highlight,
                                            unsigned char *colors, int nc_comps)
 {
   // Get a point sprite for the shape
-  vtkImageData *sprite = this->GetMarker(shape, this->Pen->GetWidth(),
+  vtkImageData *sprite = this->GetMarker(shape, static_cast<int>(this->Pen->GetWidth()),
                                          highlight);
   this->DrawPointSprites(sprite, points, n, colors, nc_comps);
 }
